requirements.txt
astroid==2.2.5
Django==2.2.3
nexmo
django-cors-headers==3.0.2
django-templated-mail==1.1.1
djangorestframework==3.9.4
djoser==1.7.0
Faker==1.0.7
fcm-django==0.2.21
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pep8==1.7.1
Pillow==6.1.0
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.1
qrcode==6.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
typed-ast==1.4.0
wrapt==1.11.2
django-filter==2.2.0
django-heroku
gunicorn
django-storages
whitenoise
boto3
wallet-py3k
